+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 # 1.0
 
  - lots of tests in examples/ and tests/
- - Tribe RISC-V CPU is developed in tribe/ folder and has it's own tests/ and linux/ folders for running Linux demo in simulator
+ - Tribe RISC-V CPU is developed in tribe_cpu/ folder and has it's own tests/ and linux/ folders for running Linux demo in simulator
  - experimental backwards SV->CPP conversion tool hdlcpp/
  - wrote documents "best_practice" and "tribe_cpu"
  - added CHANGELOG.md